Fixed bug where the generator method increased energy without requesting for the fuel.
Also fixed the bug that the shields method did not request energy to function.

+++ CrewDragonHMI/EnergyModule.cs
@@ -45,7 +45,7 @@ namespace CrewDragonHMI
             batteryLevel = level;
         }
 
-        public static bool requestEnergy(int energyRequested)
+        public static bool requestEnergy(float energyRequested)
         {
             float newBatteryLevel = getBatteryLevel() - energyRequested;
